Update sas-expiration-policy.md
The feature puts a check on specifying a maximum window for which SAS can been created, although we can still create a SAS with longer expiry.

If there is a SAS which breaches the policy of  max window specified, it will help mapping to those instances only as Policy violation. If a SAS gets expired, the calls are expected to fails. However, this feature won't be helpful in reading the expiry value of the expiry parameter and notify CX. CX need to track the expiry manually at their end.

Proposing suggestion to remove  that part. Kindly review