Update lock-resources.md
In Microsoft Sentinel UEBA - there is a reference to not having any resource locks applied to the workspace.  The link refers back to this page but I have not been able to find any documentation on this page about Microsoft Sentinel, UEBA and Resource Locks.  Here is the url https://docs.microsoft.com/en-us/azure/sentinel/enable-entity-behavior-analytics and here is what it says: Your workspace must not have any Azure resource locks applied to it. Learn more about Azure resource locking.  It needs to be prominently put on the Resource Locks page as something not to do if you would like to enable UEBA.