Update nat-gateway-integration.md
The point about using NAT GW with a subnet that has Service Endpoints is not addressed (as far as I can see) anywhere else in the docs. Fast Track for Azure Customer recently queried this as they wish to utilise both with App Service + Regional VNet integration.

The NAT Gateway will be leveraged for traffic hitting the 0.0.0.0/0 route injected by the service. Since service endpoints inject a more specific route, traffic to the service endpoint will not hit the 0.0.0.0/0 and hence bypass the NAT GW