Updates to AOAG Tutorial
This article points to an outdated link for configuring and managing quorum which I have updated to the most recent. I also added a note regarding the potential need to update the quorum votes depending on the configuration and linked to the guidance specific to SQL Server on Azure VMs. This should resolve issues that I have encountered with customers misconfiguring their quorum settings when following this tutorial.