Link to actual latest ntttcp release
The link to download ntttcp points to what was the latest release at the time, but fixes have been made to ntttcp since then. Use github's special "latest" link so that the link doesn't become stale.

There is also an irrelevant link to the toplevel github page, which I'm sure people will be able to find from the release page if they are interested. I've removed it.