Update private-endpoint-overview.md
Updated the section that talks about when creating a private endpoint to mention that a NIC is AUTOMATICALLY created for you as a result of creating the private endpoint.  The original verbiage was not clear that the additional NIC resource gets created automatically.  This can lead to confusion when deploying a private endpoint manually through the portal or through IaC.