[D365BC] Add note about the preview audience limited capabilities
Business Central doesn't support installing preview apps in customer environments. At the preview stage, only the offer listing in the marketplace can be done. This behavior is documented in Business Central docs: https://docs.microsoft.com/en-us/dynamics365/business-central/dev-itpro/developer/devenv-checklist-submission-faq#when-are-my-apps-ready-to-be-installed-in-my-business-central-environment

For visibility, it should be mentioned here too.