Update cisco-webex-tutorial.md
Evening all! 

Was working with a customer implementation today and ran into an "Invalid SAML Assertion (13)" error when attempting to follow this documentation. This particular customer is still federated (ADFS on-premises) using Citrix NetScaler as an ADFS proxy. After troubleshooting further, it was identified that we could update AuthnContextClassRef with some adjustments based on the Cisco Webex support article: https://help.webex.com/en-us/article/WBX54373/Error:-'User-Authentication-Failed,-Reason:-Invalid-SAML-Assertion-(13)'. This helped to achieve successful in application, in browser, and during the staged rollout of cloud authentication.

I feel this update would help out a LOT of customers and folks doing this same conversion work both now, and in the future! Thanks for the consideration!