(AzureCXP) MicrosoftDocs/azure-docs#88007
Edited line 231 to provide clarity on the available values of action in Firewall log as below

From : |action     |  Action taken on the request. Available values are Blocked and Allowed (for custom rules), Matched (when a rule matches a part of the request), and Detected and Blocked (these are both for mandatory rules, depending on if the WAF is in detection or prevention mode).      |

To : |action     |  Action taken on the request. Available values are: </br>**Blocked and Allowed** (for custom rules) </br>**Matched** (when a rule matches a part of the request) </br>**Detected and Blocked** (these are both for mandatory rules, depending on if the WAF is in detection or prevention mode). 

Fixes MicrosoftDocs/azure-docs#88007