[line 109] the switch for truncate is --size, not - size, also there should be a space between two parameters. remove supoernumerary hyppen character in "- copy" in an English sentense.
[line 110]
Dollar sign is not part of the command and should be removed from the beginning of the command.
It also makes it more aligned with the rest of the document.

[lines 208-212]
The correct command is useradd, not useadd.
Bash commands are case sensitive. I fixed it for useradd, passwd, su and cp commands.
define and use variable, instead of repeating storage account name several times.