Fix minor wording issues.
In the Linux output, "SLAVE" was changed to "SLA\/E" to get past submission filters.  We really should change this back, as SLAVE is what Linux outputs, notwithstanding the valid desire to not use "slave" in a context like this one.  Also fixed a wording problem in the last sentence that was introduced during earlier editing.