Update how-to-configure-rhel-centos-7.md
The documentation suggested a wrong path to Speech SDK libraries for RHEL/CentOS 7 which means users will see errors when trying to use the Speech SDK on those systems. This change fixes the path in the enviroment setting example and adds a note about the different libraries for RHEL/CentOS 7 vs. other supported Linux distributions.