(AzureCXP) fixes 3.X extensions.json link
The Extension Bundle table available here points to version 3.4.0 of extensions.json. However, the latest 3.X version of extension bundle released is 3.3.0 based on the release link here: https://github.com/Azure/azure-functions-extension-bundles/releases

This is causing confusion as the 3.X extensions.json file linked in this doc mentions that webpubsub bindings are available, but this version (3.4.0) is not available for use by Functions Core Tools as it is not released.