Update to cross-region-restore section
Added an additional bullet with an example in the note section that calls out some specifics regarding Cross Region Restores and the storage account location.
The documentation was not very clear on where the storage account needed to be located in order to show up in the drop-down menu when performing a cross region restore.  Hopefully this small addition will provide better clarity for future users of this restore capability.