Add hint on how to handle processing dead letters
The C# SDK contains a `EventGridEvent.Parse(BinaryData ...)` as well as a `EventGridEvent.ParseMany(BinaryData ...)` and it was surprising to learn that `ParseMany` existed only after having Parse fail and error message mentioning `ParseMany`.

Adding this note but I kind of wish `Parse(...)` in the SDK was simply deprecated and later removed. You'd have to know up front how many events are in the BinaryData before calling `Parse` OR you have to try a call with `Parse` catch the error and then try with `ParseMany` when it's actually simpler code to just never use `Parse` and only use `ParseMany` even if only one event is in the dead letter blob.