Update investigate-large-datasets.md
I replaced analytics with deeper analysis. I think you meant analytics as in analysis, and not analytics rules. Analytics only support looking back on 14 days, so it is not true that you can use Restore to run analytics rules. I would change the wording to avoid confusing because initially I thought this referred to analytics rules.