Update create-file-share doc.
Adding a line in "how to set up a storage account" section to include a warning about the lenght of the storage account name parameter. 
Because if it's longer than 15 characters the domain join will fail as currently a bigger length is not supported to create an AD object for the storage account. 
This is a small detail but it is impacting customer experience has the customer might create a storage account with a name with more than 15 letters and then this storage account is no good and they need to redo all the creation steps.