Update azure-monitor-costs.md
I work in Azure Monitor and I'm responsible for the "horizontals" project which is standardizing Azure Monitor information across all docsets. 

This is a great article. We noticed it because we are rewriting our Azure monitor cost article, which is a mess. :/.  We may model our changes on your article.  

That said...The current title of this article is incorrect from an H1 standpoint in that it does not match the title in the meta data. I've changed both to give a proper context


-------
cc: @Heidilohr @brwen