Update azure-app-offer-listing.md
To maximise the accessibility of the marketplace experience, this update suggests that offer creators consider the accessibility of added media.  As this uploaded content will sit alongside Microsoft owned content in the final customer marketplace navigation and evaluation experience, this tip is a nudge with a link to Microsoft public facing content on how to create accessible media.