Note inaccurate speed reporting on ENDURE
ENDURE SKUs don't reliably report the speed of their InfiniBand connections via the synthetic NIC that's surfaced in the Guest OS. Whenever it doesn't match the expected numbers for the VM size, it causes confusion. Adding a note to the docs page for H-series VMs to clarify that this does not translate to an actual performance issue.