Include note about immutability of CosmosClientOptions properties
There is some potentially misleading behavior around the CosmosClientOptions class. Despite having a setter, changes to the properties are not respected and will silently do nothing. This was not mentioned in the documentation anywhere when I was learning about bulk support, and I think it could help someone out in the future.

Related to https://github.com/Azure/azure-cosmos-dotnet-v3/issues/3012 and https://github.com/Azure/azure-cosmos-dotnet-v3/issues/3091