Add missing custom role
In order to do Azure CNI networking you also require the permission Microsoft.Authorization/roleAssignments/write. This is documented in the Azure portal when you attempt to create a cluster with Azure ANI and a new virtual network - if the permission is missing, the portal displays a warning stating it is required.