Update gateway.md
I verified (and shared with Shayoni) that the LA Gateway only works with AMA if the source server is also running the same agent. You cannot have MMA on one server and AMA on another server. I verified this for Windows, so I added it specifically for Windows. It should also be verified for Linux.