Update lifecycle-management-overview.md
A lot many times there is ask from the customer at which time the Lifecycle management policy will be executing. There is no way to track that as it is a background scheduling process however the platform will be running it once a day. Hence, proposing this as part of FAQ. Please review