Outdated information - "|Display password expiry"
The following information is outdated - 

|Display password expiry notifications in the Office Portal and on the Windows 10 desktop.| | |![Recommended](./media/whatis-hybrid-identity/ic195031.png)| 

1.
"Display password expiry notifications in the Office Portal" is no longer supported per following document - 
https://docs.microsoft.com/en-us/microsoft-365/admin/manage/set-password-expiration-policy?view=o365-worldwide
---------------------
 Important
Password expiration notifications are no longer supported in Office web apps or the admin center.
---------------------

2. 
"Display password expiry notifications on the Windows 10 desktop" is done through GPO of on-premises ADDS so it seems irrelevant to this context of hybrid identity. Also, when password is not sync'ed to Azure AD, the sync'ed user cannot sign into Azure AD Joined devices, so there doesn't seem to be any difference between the user being on-prem only user or hybrid ID. Suggest removing this line "Display password expiry notifications in the Office Portal" from the table for clarity.