Update index.yml
Remove all references to Oracle workloads on BareMetal Infrastructure.  The team which created this page and added references to BMI/Oracle is no longer at Microsoft.  This page must be edited to remove all references to avoid confusion for customers.  If this service is ever offered again in future, then the necessary verbiage can be restored by that team.  Until then, there is no team, so the doc page must reflect the reality.