Update connect-install-beeline.md
Beeline is a client tool. We don't "connect to Beeline" but we do connect to "Hive" using Beeline. So, correcting the mistake in the title as well as in the 1st paragraph.
Also noticed a typo in : "From cluster head or inside Azure Virtual Network with Apache Spark" . We should use "From cluster head *node* or inside Azure Virtual Network with Apache Spark" instead. Corrected this too.