Update send-email-net.md
Hey Bala,

I included tables for the object model descriptions and message statuses. I also made a couple of changes where Sms was still referenced. I'll add in the wrapper for this include, as well. Thank you for doing so much of this work - you REALLY helped me out! Have a great weekend!