Update entitlement-management-organization.md
Hi,

I have worked on couple of support cases where customers were reporting an issue with users coming from a connected org requesting an access package and it results in a new suggested connected org. This is because the user is using their personal account instead of the organization account. The connected org represents the Azure AD version of the organization so the portal did not identify the user coming from this tenant since they are using their personal account hence attempting to create a suggested connected org. I have added a note box in the doc.