Update to reflect ability to pull NPSP data schema
The purpose of this change is to call out the ability of this connector to pull many different data schemas in Salesforce, specifically the nonprofit success pack.  We have worked with both Salesforce and the Azure Data Factory team to validate this data pull success for this schema.  Calling out this ability is critical for the MS Tech for Social Impact Sales team.