Add note to semantic type example
It was reported by ADT Explorer users that a semantic property defined in one of their models was missing on the twin-level properties window. This was caused because they were not following the sequential order shown in the example. In the proposed change request, a note was added to the Semantic type example section highlighting the fact that the order of the elements in the array is relevant.