Updating the Azure IP Datacenter link with latest
The link in line number 46 is referencing to the deprecating data center IP addresses for Azure. the latest is available in the link https://www.microsoft.com/en-us/download/confirmation.aspx?id=56519 and documented currently as Azure IP Ranges and Service Tags – Public Cloud. FIXED the link to highlight the same based on a recent customer query. 
#ATCP.