Fixed command to verify service principle role
when creating the roles we are scoping to the resource group, not the subscription. The original command only got the roles for the subscription, so we would never see the roles we just created. Added the -scope parameter to the command so that we can confirm the roles are created.
az role assignment list --scope subscriptions/<SubscriptionID>/resourceGroups/<resourcegroup> -o table