Added clarification on validation technical profile execution
Ran into an issue recently where we wanted to call some validation technical profiles to take advantage of the ContinueOnError attribute of a ValidationTechnicalProfile element. The documentation didn't make it immediately clear (to me) that there must be some input from the user in the calling self-asserted technical profile to execute, and B2C does not alert the user to why it skipped the validation technical profiles.

This update should make this clear and save someone else some time.