Update live-stream.md
Supportability PM Vitaly reported that customers had said that sections on using unsecure control channel was unclear. Updated section. However, there is a screen shot that included in that section that is kind of old (./media/live-stream/live-stream-auth.png). I tried to update that as well, but I don't have push permissions to the repository. Sent Email to Aaron Maxwell with image attached.