Update develop-tables-cetas.md
Gen2 Storage Account, which is a Synapse prerequesite -> https://docs.microsoft.com/en-us/sql/t-sql/statements/create-external-data-source-transact-sql?view=azure-sqldw-latest&tabs=dedicated
The doc should be pointing to .dfs. endpoint instead of blob. I was trying to demo with this example and it was failing. Please note this is also documented here: https://docs.microsoft.com/en-us/answers/questions/825946/external-table-location-path-is-not-valid-location.html