Azure NetApp Files Global Vnet Peering
As of my tests in my personal and client's environment, we have a hub and spoke topology, the hub is in west europe and the spoke in east us, when activating Standard Network Features and forwarding the traffic to Azure Firewall connectivity using Global Vnet Peering is working. Documentation should reflect this. Thanks.