AVS to On-premises Visual in this article is mis-represented
After consulting with one of our Networking GBBs, the AVS ER circuit should be connected to the ER GW in (existing) Hub VNet, same ER GW as where the on-premises ER circuit is connected.  Non-transitive by default, until the ER Global Reach feature is enabled to allow route exchange.  With both ER circuits connected to the ER Gateway with Global Reach feature, AVS and on-premises can communicate.  Please update the Visual embedded in this article.