Update workload-identity-federation-create-trust-gcp.md
Customer was getting AADSTS50027 when using below and the issue was resolved by using the code provided below as what worked. Please review and update the 
doc as necessary.

Code that was giving AADSTS50027

Cx is following the document: https://docs.microsoft.com/en-us/azure/active-directory/develop/workload-identity-federation-create-trust-gcp?tabs=typescript

And finds issue with part of the code: this seems to be incorrect.  As ClientAssertion is passed as string.
 return getGoogleIDToken() // calling this directly just for clarity, 
                               // this should be a callback
        // pass this as a client assertion to the confidential client app
        .then((clientAssertion:any)=> {
            var msalApp: any;
            msalApp = new msal.ConfidentialClientApplication({
                auth: {
                    clientId: this.clientID,
                    authority: this.aadAuthority + this.tenantID,
                    clientAssertion: clientAssertion,
                }
            });
            return msalApp.acquireTokenByClientCredential({ scopes })
        })

what worked


The example code has that the clientAssertion will be the output of the getGoogleIDToken(). The getGoogleIDToken returns either a Promise or a response object. To make it work the getGoogleIDToken needs to return the payload of the request which will be the token itself.

instead of return fetch(endpoint, options) should I used Axios (haven't tried yet with fetch)

what worked for me
Text
let aadAudience = "api://AzureADTokenExchange"
    const jwt = axios({
            url: "http://metadata.google.internal/computeMetadata/v1/instance/service-accounts/default/identity?audience=" + aadAudience,
            method: "GET",
            headers: {
                    "Metadata-Flavor": "Google"
            }}).then(response => {
                    console.log("AXIOS RESPONSE");
                    return response.data;
                   });
    return jwt;
 





#ATCP