Update sql-data-warehouse-tables-overview.md
Added a new join criteria to the dm_pdw_nodes_db_partition_stats table in the vTableSizes view definition (AND i.[index_id] = nps.[index_id]). Without this records counts are inflated because the query is multiplying the number of records by the number of secondary indexes on the table. This change ensures only the primary index/storage is considered for counting.