Update fqdn-tags.md
In the documentation the tags had spaces in them but when using them they don't have spaces. It's not a problem if it's click-ops in the portal but when using IaC it leads to failures in deployment because of the spaces when cut and pasting the tags out of the doc.