Update Subscription PS Commands
The documentation references output from $sub.TenantId and $sub.SubscriptionId in the "Where to Obtain" section of the table under "Define Packer Template", however there are no commands specified that would create/define the $sub variable and its values. Updated the PowerShell example command for Get-AzSubscription to properly store results in variable $sub so that a user can properly retrieve the Tenant and Sub ID values.