Fix typo in nodepool snapshot creation
This is a simple PR to fix the creation of a nodepool snapshot command line.

If we use `az aks snapshot create` as documented previously, the cli actually tries to create a full snapshot of the cluster, instead of just the nodepool. Also, the command line fails with message: `the following arguments are required: --cluster-id`.

The fix is simple, we just need to use `az aks nodepool snapshot create` instead.