Add note about MapControllers in ASP.NET Core
MapControllers is not called by default when starting with a new ASP.NET Core Web App that uses Razor Pages. However, since the Microsoft Identity UI uses it, forgetting to add `app.MapControllers()` will cause sign-in and sign-out links to stay blank and manually navigating to `/MicrosoftIdentity/Account/SignOut` to result in a 404.

Considering the documentation already mentions `UseAuthentication` and `AddRazorPages` in the samples, I feel like `MapControllers` should be added as well.