Max Clusters typo - correct value should be 12
"The minimum number of hosts per cluster is three. More hosts can be added one at a time, up to a maximum of 16 hosts per cluster. The maximum number of clusters per private cloud is four"

The value needs to be updated to twelve which is the actual limit. See here:
https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/azure-subscription-service-limits
https://docs.microsoft.com/en-us/azure/azure-vmware/concepts-private-clouds-clusters#host-maintenance-and-lifecycle-management