Update URL for Personal Accounts
Issue: Current URL for personal accounts mentioned on the document which is https://myapps.microsoft.com does not works any longer. The URL simply returns an error stating: "You can only access this site with a work account. Sign out to log in with another account." The only way to access a particular organization using MSA/Personal Account is to use a direct link with specified tenant id. https://myaccount.microsoft.com?tenantid=contoso.onmicrosoft.com

Ask: Update correct URL on the document as mentioned in above issue statement and on the change request. MSA users follow the document and fail to login to https://myapps.microsoft.com/ with error "You can only access this site with a work account. Sign out to log in with another account." This directly impacts customer experience as the workflow on the documentation does not works and impacts CSAT for CSS/CXP cases.

Impact: Updating this document would be able increase deflection as document would be able to provide appropriate data to customers. Currently MSA users are unaware about this change and would result in following the document and creating additional CSS cases.

This has helped us fix the issue on 2206210040001442. ADO Request for Internal Tracking: https://msazure.visualstudio.com/One/_workitems/edit/14892527