Clarification about indexers behaviors
There have been already 2-3 support cases in the past month or so, complaining that their indexer is idle and when raising to developers, it would be expected, if there are other jobs in queue.  Since indexers do not have an SLA, also adding there is no predictability, to avoid customers having wrong expectations.