Update cache-reserved-pricing.md
This explanation "Reserved capacity is sold in increments of nodes. Each shard contains 2 nodes by default, so to buy reserved capacity for a shard buy 2 reserved capacity." is referred on the reservation purchasing page of Azure Portal. 

This request is because it is difficult for new users to know the number of nodes (2 nodes) by default.