Adding note to env secret yaml example
This yaml example includes both an env variable referencing a secret and a separate volume/volumeMount for the same secret. Adding in a note to assert that this is for illustrative purposes as customer might think that these two methods are dependent of one another - this is not the case - these methods can be used interpedently of the other for the purpose of getting the same secret.