Update deploy.md
There is an incorrect information in the documentation. "In the designer pane, select **Function-Try** to configure the target settings". That won't configure anything and people may stuck. The proper information should be "In the designer pane, select **Function-Try** to configure the target settings and click the "</> Code view" button at the top menu to edit the code for the "Function-Try" element.