Update cache-how-to-premium-persistence.md
Firewall Exception on the storage account will affect the persistence process. 
If using Managed Identity in order to authenticate the storage account, this shall help manage firewall exceptions and if not and key being used then the process may tend to break. This seem to have been there previously but not now and hence proposing to have this as an FAQ. Please review