Issue when mounting file share of type NFS (BYOS) to App Service
An email was sent to the "Azure Web Apps (Antares) discussions" alias requesting this change.  Azure Files are supported for Linux App Service.  I think the nuance here is that we only support (currently) Azure Files SMB, but not Azure Files NFS.  Valid point though to call this out explicitly in our docs and such.