Clarify requirements for data used in custom model training
Through discussion with the product team I have confirmed that upon training and analysis a "single instance" of the document/form should be used as input data. So for example, "concatenating" several instances of the form during API calls, for convenience's sake, is not supported and would cause unexpected results.
This edit is to modify the documentations pertaining to training practices.