Remove Windows 7 as it is no longer supported
Windows 7 is no longer a supported operating system. This is also outlined in the documentation for implementing HAADJ for down level devices. 
https://docs.microsoft.com/en-us/azure/active-directory/devices/howto-hybrid-join-downlevel

If we don't plan to support customers creating tickets for HAADJ issues with Windows 7 then it should no longer list Windows 7 as a supported OS in this documentation. 

Other documentation referencing the end of support day for Windows 7.
https://support.microsoft.com/en-us/windows/windows-7-support-ended-on-january-14-2020-b75d4580-2cc7-895a-2c9c-1466d9a53962
https://www.microsoft.com/en-us/windows/end-of-support