Remove line and an incomplete sentence.
Line 20  didn't add any value and also included an incomplete sentence:
SQL API SDKs are available for a wide variety of platforms and programming languages. If you haven't worked 

The next line basically states the same thing so there isn't a need to repeat the info.  It also looks like the thought moved to the next paragraph anyway "If you haven't worked with one of these SDKs before, see the *"Quickstart"* article for the appropriate SDK:"