Added CA Certificate Note
Based on recent customer issue, the current behavior of Self-Hosted gateway is that it does not present all intermediate certificates in the chain to the client connecting to the gateway. This can cause SSL handshake failure for clients depending on the certs being presented. Managed gateway already presents interim certs to the clients. We have a bug open to address this in self-hosted gateway, but until thats fixed, we need add this documentation.