Manual setup not working for AES256 ActiveDirectorySamAccountName ends with a '$' sign
Manual setup is used when AD users are not synced with AAD (the AzFilesHybrid module requires the user running the script to be synced with AAD)
The manual setup instructions are not working - Its needs to specify that the **ActiveDirectorySamAccountName** property must be provided without the trailing '$' sign, otherwise AES256 encryption will not work. When using the AzFilesHybrid module, the '$' sign is removed by the Set-StorageAccountDomainProperties function.
Also - the setspn command is missing the computer name at the end.