Increase visibility of limitations
The limitations should not be included in the introduction, as it is easy to miss. The introduction typically includes a description of the article, and should not contain key information relating to support of the product. 

This edit produces a similar structure to https://docs.microsoft.com/en-us/azure/app-service/overview-vnet-integration#limitations.