inaccurate results by design
This is modified based on internal feedback. 

Comment - 
"Customer explain that the documentation is not clear enough to warn customers about this by design behavior, they asked of we can adjust the documentation so that it appears as warning, or important note, …etc. "

This is more likely require a "CAUTION" but i have created this as "IMPORTANT". 
Please feel free to replace "CAUTION" is required. 

> [!CAUTION]  
> After creating a table with primary key or unique constraint in dedicated SQL pool, users need to make sure all values in those columns are unique.
> A violation of that may cause the query to return inaccurate result.