Update validate-jwt policy for RS256 signing algorithms
The pfx file in general contains a public key as well as its corresponding private key. In this case of jwt validation for RS256 signing algorithms, we need only the public key of the certificate for validation. But APIM custom client certificate uploads don't support cert format and it supports only pfx format. So we have to create a pfx file only with the public key for custom client certificate upload. So it's better to call out this format explicitly in the document as generally, a pfx format has a private key as well. This clarifies to the user that we use should only pfx file but with only the public key and it also clarifies the security concerns about uploading a pfx file.